requirements.txt
python-telegram-bot==21.5
aiohttp==3.9.5
apscheduler==3.10.4
pandas==2.2.2
openpyxl==3.1.5
pytz==2024.1
